+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/core/DecodeTransferMetadataServletFilter.java
@@ -21,8 +21,11 @@ package com.tencent.cloud.metadata.core;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
+import java.util.Arrays;
+import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
@@ -37,12 +40,15 @@ import org.slf4j.LoggerFactory;
 
 import org.springframework.core.annotation.Order;
 import org.springframework.lang.NonNull;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import static com.tencent.cloud.common.constant.ContextConstant.UTF_8;
 import static com.tencent.cloud.common.constant.MetadataConstant.HeaderName.CUSTOM_DISPOSABLE_METADATA;
 import static com.tencent.cloud.common.constant.MetadataConstant.HeaderName.CUSTOM_METADATA;
+import static com.tencent.cloud.common.metadata.MetadataContext.FRAGMENT_RAW_TRANSHEADERS;
+import static com.tencent.cloud.common.metadata.MetadataContext.FRAGMENT_RAW_TRANSHEADERS_KV;
 
 /**
  * Filter used for storing the metadata from upstream temporarily when web application is
@@ -71,7 +77,33 @@ public class DecodeTransferMetadataServletFilter extends OncePerRequestFilter {
 
 		MetadataContextHolder.init(mergedTransitiveMetadata, mergedDisposableMetadata);
 
+		setCompleteTransHeaderIntoMC(httpServletRequest);
 		filterChain.doFilter(httpServletRequest, httpServletResponse);
+	}
+
+	/**
+	 * According to HttpServletRequest and trans-headers(key list in string type) in metadata, build
+	 * the complete headers(key-value list in map type) into metadata.
+	 */
+	private void setCompleteTransHeaderIntoMC(HttpServletRequest httpServletRequest) {
+		// transHeaderMetadata: for example, {"trans-headers" : {"header1;header2;header3":""}}
+		Map<String, String> transHeaderMetadata =  MetadataContextHolder.get()
+				.getFragmentContext(FRAGMENT_RAW_TRANSHEADERS);
+		if (!CollectionUtils.isEmpty(transHeaderMetadata)) {
+			Optional<String> transHeaders = transHeaderMetadata.keySet().stream().findFirst();
+			String[] transHeaderArray = transHeaders.get().split(";");
+			Enumeration<String> httpHeaders = httpServletRequest.getHeaderNames();
+			while (httpHeaders.hasMoreElements()) {
+				String httpHeader = httpHeaders.nextElement();
+				Arrays.stream(transHeaderArray).forEach(transHeader -> {
+					if (transHeader.equals(httpHeader)) {
+						String httpHeaderValue = httpServletRequest.getHeader(httpHeader);
+						MetadataContextHolder.get().putContext(FRAGMENT_RAW_TRANSHEADERS_KV,	httpHeader, httpHeaderValue);
+						return;
+					}
+				});
+			}
+		}
 	}
 
 	private Map<String, String> getInternalMetadata(HttpServletRequest httpServletRequest, String headerName) {